chore: Split the class FlinkLanguageServiceProxy into multiple files
