+++ src/main/java/io/confluent/idesidecar/websocket/messages/FlinkLanguageServiceAuthMessage.java
@@ -1,0 +1,10 @@
+package io.confluent.idesidecar.websocket.messages;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record FlinkLanguageServiceAuthMessage (
+    @JsonProperty(value = "Token") String token,
+    @JsonProperty(value = "EnvironmentId") String environmentId,
+    @JsonProperty(value = "OrganizationId") String organizationId
+) {
+}
