+++ src/main/java/io/confluent/idesidecar/websocket/resources/FlinkLanguageServiceProxy.java
@@ -1,164 +1,32 @@
 package io.confluent.idesidecar.websocket.resources;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import io.confluent.idesidecar.restapi.connections.CCloudConnectionState;
 import io.confluent.idesidecar.restapi.connections.ConnectionStateManager;
 import io.confluent.idesidecar.restapi.exceptions.ConnectionNotFoundException;
+import io.confluent.idesidecar.websocket.proxy.FlinkLanguageServiceProxyClient;
+import io.confluent.idesidecar.websocket.proxy.ProxyContext;
 import io.quarkus.logging.Log;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-import jakarta.websocket.ClientEndpoint;
 import jakarta.websocket.CloseReason;
 import jakarta.websocket.CloseReason.CloseCodes;
-import jakarta.websocket.ContainerProvider;
 import jakarta.websocket.OnClose;
 import jakarta.websocket.OnMessage;
 import jakarta.websocket.OnOpen;
 import jakarta.websocket.Session;
 import jakarta.websocket.server.ServerEndpoint;
 import java.io.IOException;
-import java.net.URI;
-import java.util.concurrent.Future;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicInteger;
-import org.eclipse.microprofile.config.ConfigProvider;
 
 @ServerEndpoint("/flsp")
 @ApplicationScoped
 public class FlinkLanguageServiceProxy {
 
-  static final String LANGUAGE_SERVICE_URL_PATTERN = ConfigProvider
-      .getConfig()
-      .getValue("ide-sidecar.flink-language-service-proxy.url-pattern", String.class);
-  static final Integer MAX_RECONNECT_ATTEMPTS = ConfigProvider
-      .getConfig()
-      .getValue("ide-sidecar.flink-language-service-proxy.reconnect-attempts", Integer.class);
-  static final String CONNECTION_ID_PARAM_NAME = "connectionId";
-  static final String REGION_PARAM_NAME = "region";
-  static final String PROVIDER_PARAM_NAME = "provider";
-  static final String ENVIRONMENT_ID_PARAM_NAME = "environmentId";
-  static final String ORGANIZATION_ID_PARAM_NAME = "organizationId";
-
-  static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
-
-  record AuthMessage (
-      @JsonProperty(value = "Token") String token,
-      @JsonProperty(value = "EnvironmentId") String environmentId,
-      @JsonProperty(value = "OrganizationId") String organizationId
-  ) {
-  }
-
-  record ProxyContext (
-      String connectionId,
-      String region,
-      String provider,
-      String environmentId,
-      String organizationId
-  ) {
-    static ProxyContext from(Session session) {
-      var paramMap = session.getRequestParameterMap();
-      return new ProxyContext(
-          paramMap.get(CONNECTION_ID_PARAM_NAME).get(0),
-          paramMap.get(REGION_PARAM_NAME).get(0),
-          paramMap.get(PROVIDER_PARAM_NAME).get(0),
-          paramMap.get(ENVIRONMENT_ID_PARAM_NAME).get(0),
-          paramMap.get(ORGANIZATION_ID_PARAM_NAME).get(0)
-      );
-    }
-
-    String getConnectUrl() {
-      // TODO: I guess this won't work for private networks, we'll need something more sophisticated
-      return LANGUAGE_SERVICE_URL_PATTERN
-          .replace("{{ region }}", region)
-          .replace("{{ provider }}", provider);
-    }
-  }
-
-  @ClientEndpoint
-  class ProxyClient {
-    Session remoteSession;
-    Session localSession;
-    ProxyContext context;
-    AtomicInteger reconnectAttempts = new AtomicInteger(0);
-
-    public ProxyClient(ProxyContext context, Session localSession) {
-      this.context = context;
-      this.localSession = localSession;
-      try {
-        var container = ContainerProvider.getWebSocketContainer();
-        container.connectToServer(this, URI.create(context.getConnectUrl()));
-      } catch (Exception e) {
-        throw new RuntimeException(e);
-      }
-    }
-
-    @OnOpen
-    public void onOpen(Session remoteSession) {
-      this.remoteSession = remoteSession;
-      try {
-        var connection = (CCloudConnectionState) connectionStateManager.getConnectionState(
-            context.connectionId);
-        this.remoteSession.getAsyncRemote().sendText(
-            OBJECT_MAPPER.writeValueAsString(
-                new AuthMessage(
-                    connection.getOauthContext().getDataPlaneToken().token(),
-                    context.environmentId,
-                    context.organizationId
-                )
-            )
-        );
-      } catch (Exception e) {
-        Log.errorf("Failed to send initial auth message: %s", e.getMessage());
-      }
-    }
-
-    @OnMessage
-    public void onMessage(String message) {
-      localSession.getAsyncRemote().sendText(message);
-      // Connection seems to be healthy, let's reset the number of reconnect attempts
-      reconnectAttempts.set(0);
-    }
-
-    @OnClose
-    public void onClose(Session session) throws IOException {
-      // Increase number of reconnect attempts and close the session if the maximum number of
-      // reconnect attempts has been reached
-      if (reconnectAttempts.incrementAndGet() > MAX_RECONNECT_ATTEMPTS) {
-        Log.errorf("Max reconnect attempts reached. Closing session.");
-        localSession.close(
-            new CloseReason(CloseCodes.CLOSED_ABNORMALLY, "Max reconnect attempts reached.")
-        );
-        return;
-      }
-
-      this.remoteSession = null;
-      try {
-        var container = ContainerProvider.getWebSocketContainer();
-        container.connectToServer(this, URI.create(context.getConnectUrl()));
-      } catch (Exception e) {
-        throw new RuntimeException(e);
-      }
-    }
-
-    public Future<Void> sendToCCloud(String message) {
-      return this.remoteSession.getAsyncRemote().sendText(message);
-    }
-
-    public void close() {
-      try {
-        remoteSession.close();
-      } catch (IOException e) {
-        Log.error("Could not close WebSockets session to CCloud Language Service.", e);
-      }
-    }
-  }
-
   @Inject
   ConnectionStateManager connectionStateManager;
 
-  Map<String, ProxyClient> proxyClients = new ConcurrentHashMap<>();
+  Map<String, FlinkLanguageServiceProxyClient> proxyClients = new ConcurrentHashMap<>();
 
   @OnOpen
   public void onOpen(Session session) throws IOException {
@@ -170,7 +38,7 @@ public class FlinkLanguageServiceProxy {
       return;
     }
     try {
-      var connection = connectionStateManager.getConnectionState(context.connectionId);
+      var connection = connectionStateManager.getConnectionState(context.connectionId());
       if (connection instanceof CCloudConnectionState cCloudConnectionState) {
         if (cCloudConnectionState.getOauthContext() == null
             || cCloudConnectionState.getOauthContext().hasReachedEndOfLifetime()
@@ -179,11 +47,11 @@ public class FlinkLanguageServiceProxy {
               new CloseReason(
                   CloseCodes.CANNOT_ACCEPT,
                   "Connection with ID=%s does not have a data plane token.".formatted(
-                      context.connectionId)
+                      context.connectionId())
               )
           );
         } else {
-          var client = new ProxyClient(context, session);
+          var client = new FlinkLanguageServiceProxyClient(context, session, connectionStateManager);
           proxyClients.put(session.getId(), client);
           Log.infof("Added LSP client for session ID=%s", session.getId());
         }
@@ -191,7 +59,7 @@ public class FlinkLanguageServiceProxy {
         session.close(
             new CloseReason(
                 CloseCodes.CANNOT_ACCEPT,
-                "Connection with ID=%s is not of type CCLOUD.".formatted(context.connectionId)
+                "Connection with ID=%s is not of type CCLOUD.".formatted(context.connectionId())
             )
         );
       }
@@ -199,7 +67,7 @@ public class FlinkLanguageServiceProxy {
       session.close(
           new CloseReason(
               CloseCodes.CANNOT_ACCEPT,
-              "Could not find connection with ID=%s.".formatted(context.connectionId)
+              "Could not find connection with ID=%s.".formatted(context.connectionId())
           )
       );
     }
